+++ t/apicast-policy-token-introspection.t
@@ -1,6 +1,10 @@
 use lib 't';
 use Test::APIcast::Blackbox 'no_plan';
 
+use Cwd qw(abs_path);
+
+our $rsa = `cat t/fixtures/rsa.pem`;
+
 run_tests();
 
 __DATA__
@@ -28,6 +32,7 @@ Token introspection policy check access token.
           {
             "name": "apicast.policy.token_introspection", 
             "configuration": {
+              "auth_type": "client_id+client_secret",
               "client_id": "app",
               "client_secret": "appsec",
               "introspection_url": "http://test_backend:$TEST_NGINX_SERVER_PORT/token/introspection"
@@ -77,6 +82,7 @@ Token introspection policy return "403 Unauthorized" if access token is already 
           {
             "name": "apicast.policy.token_introspection", 
             "configuration": {
+              "auth_type": "client_id+client_secret",
               "client_id": "app",
               "client_secret": "appsec",
               "introspection_url": "http://test_backend:$TEST_NGINX_SERVER_PORT/token/introspection"
@@ -126,6 +132,7 @@ Token introspection policy return "403 Unauthorized" if IdP response error statu
           {
             "name": "apicast.policy.token_introspection", 
             "configuration": {
+              "auth_type": "client_id+client_secret",
               "client_id": "app",
               "client_secret": "appsec",
               "introspection_url": "http://test_backend:$TEST_NGINX_SERVER_PORT/token/introspection"
@@ -177,6 +184,7 @@ Token introspection policy return "403 Unauthorized" if IdP response invalid con
           {
             "name": "apicast.policy.token_introspection",
             "configuration": {
+              "auth_type": "client_id+client_secret",
               "client_id": "app",
               "client_secret": "appsec",
               "introspection_url": "http://test_backend:$TEST_NGINX_SERVER_PORT/token/introspection"
@@ -205,6 +213,7 @@ Authorization: Bearer testaccesstoken
 --- error_code: 403
 --- error_log
 [error]
+
 === TEST 5: Token introspection request is failed with null response
 Token introspection policy return "403 Unauthorized" if IdP null response .
 --- backend
@@ -228,6 +237,7 @@ Token introspection policy return "403 Unauthorized" if IdP null response .
           {
             "name": "apicast.policy.token_introspection",
             "configuration": {
+              "auth_type": "client_id+client_secret",
               "client_id": "app",
               "client_secret": "appsec",
               "introspection_url": "http://test_backend:$TEST_NGINX_SERVER_PORT/token/introspection"
@@ -256,3 +266,138 @@ Authorization: Bearer testaccesstoken
 --- error_code: 403
 --- error_log
 [error]
+
+=== TEST 6: Token introspection request success with oidc issuer endpoint
+Token introspection policy retrieves client_id and client_secret and 
+introspection endpoint from the oidc_issuer_endpoint of the service configuration.
+--- backend
+  location /token/introspection {
+    content_by_lua_block {
+      local credential = ngx.decode_base64(require('ngx.re').split(ngx.req.get_headers()['Authorization'], ' ', 'oj')[2])
+      require('luassert').are.equal('app:appsec', credential)
+      ngx.say('{"active": true}')
+    }
+  }
+
+  location /transactions/oauth_authrep.xml {
+    content_by_lua_block {
+      ngx.exit(200)
+    }
+  }
+
+--- configuration
+{
+  "oidc": [
+    {
+      "issuer": "https://example.com/auth/realms/apicast",
+      "config": {
+        "public_key": "MFwwDQYJKoZIhvcNAQEBBQADSwAwSAJBALClz96cDQ965ENYMfZzG+Acu25lpx2KNpAALBQ+catCA59us7+uLY5rjQR6SOgZpCz5PJiKNAdRPDJMXSmXqM0CAwEAAQ==",
+        "openid": {
+          "id_token_signing_alg_values_supported": [ "RS256" ],
+          "token_introspection_endpoint": "http://test_backend:$TEST_NGINX_SERVER_PORT/token/introspection"
+        }
+      }
+    }
+  ],
+  "services": [
+    {
+      "id": 42,
+      "backend_version": "oauth",
+      "backend_authentication_type": "service_token",
+      "backend_authentication_value": "token-value",
+      "proxy": {
+        "authentication_method": "oidc",
+        "oidc_issuer_endpoint": "http://app:appsec@test_backend:$TEST_NGINX_SERVER_PORT/issuer/endpoint",
+        "api_backend": "http://test:$TEST_NGINX_SERVER_PORT/",
+        "proxy_rules": [
+          { "pattern": "/", "http_method": "GET", "metric_system_name": "hits", "delta": 1 }
+        ],
+        "policy_chain": [
+          {
+            "name": "apicast.policy.token_introspection",
+            "configuration": {
+              "auth_type": "use_3scale_oidc_issuer_endpoint"
+            }
+          },
+          { "name": "apicast.policy.apicast" }
+        ]
+      }
+    }
+  ]
+}
+--- upstream
+  location /echo {
+    content_by_lua_block {
+      ngx.say('yay, api backend');
+    }
+  }
+--- request
+GET /echo
+--- more_headers eval
+use Crypt::JWT qw(encode_jwt);
+my $jwt = encode_jwt(payload => {
+  aud => 'the_token_audience',
+  nbf => 0,
+  iss => 'https://example.com/auth/realms/apicast',
+  exp => time + 3600 }, key => \$::rsa, alg => 'RS256');
+"Authorization: Bearer $jwt"
+--- error_code: 200
+--- response_body
+yay, api backend
+--- no_error_log
+[error]
+oauth failed with
+
+=== TEST 7: Token introspection request fails with app_key
+Token introspection policy retrieves client_id and client_secret and 
+introspection endpoint from the oidc_issuer_endpoint of the service configuration.
+When authentication_method = 1, the request fails.
+--- backend
+  location /transactions/authrep.xml {
+    content_by_lua_block {
+      ngx.exit(200)
+    }
+  }
+
+--- configuration
+{
+  "services": [
+    {
+      "id": 42,
+      "backend_version": "1",
+      "backend_authentication_type": "service_token",
+      "backend_authentication_value": "token-value",
+      "proxy": {
+        "authentication_method": "1",
+        "oidc_issuer_endpoint": "http://app:appsec@test_backend:$TEST_NGINX_SERVER_PORT/issuer/endpoint",
+        "api_backend": "http://test:$TEST_NGINX_SERVER_PORT/",
+        "proxy_rules": [
+          { "pattern": "/", "http_method": "GET", "metric_system_name": "hits", "delta": 1 }
+        ],
+        "policy_chain": [
+          {
+            "name": "apicast.policy.token_introspection",
+            "configuration": {
+              "auth_type": "use_3scale_oidc_issuer_endpoint"
+            }
+          },
+          { "name": "apicast.policy.apicast" }
+        ]
+      }
+    }
+  ]
+}
+--- upstream
+  location /echo {
+    content_by_lua_block {
+      ngx.say('yay, api backend');
+    }
+  }
+--- request
+GET /echo?user_key=userkey
+--- error_code: 403
+--- response_body
+Authentication failed
+--- no_error_log
+[error]
+oauth failed with